iterator: extract iterator module
query_result.rs needed to import metadata structs just because iterator
methods were defined here.

I also removed the commented CassIterator methods. I don't see a purpose
of having this comment. I'll do the same for other types as well in later commit.
